+++ hamza-server/src/config.ts
@@ -1,0 +1,7 @@
+import * as process from 'node:process';
+
+class Config {
+    public static get(): Config {
+        return process.env.MASSMARKET_THING;
+    }
+}
